# backend/swagger.yaml
openapi: 3.0.0
info:
  title: API de Gestión de Candidatos
  version: 1.0.0
  description: API para gestionar candidatos en un sistema Kanban

paths:
  /positions/{id}/candidates:
    get:
      summary: Obtiene los candidatos de una posición
      description: Devuelve todos los candidatos en proceso para una posición específica
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: ID de la posición
      responses:
        '200':
          description: Lista de candidatos obtenida exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    full_name:
                      type: string
                      example: "John Doe"
                    current_interview_step:
                      type: string
                      example: "Technical Interview"
                    average_score:
                      type: number
                      format: float
                      example: 4.5
        '500':
          description: Error del servidor
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  error:
                    type: string

  /candidates/{id}/stage:
    put:
      summary: Actualiza la etapa de un candidato
      description: Actualiza la etapa del proceso de entrevista de un candidato específico
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: ID del candidato
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                new_stage:
                  type: string
                  example: "Final Interview"
              required:
                - new_stage
      responses:
        '200':
          description: Etapa actualizada exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Candidate stage updated successfully"
                  candidate_id:
                    type: integer
                    example: 456
                  new_stage:
                    type: string
                    example: "Final Interview"
        '500':
          description: Error del servidor
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Error al actualizar la etapa del candidato"
                  error:
                    type: string

  /candidates:
    post:
      summary: Añade un nuevo candidato
      description: Crea un nuevo registro de candidato con su información personal, educación y experiencia laboral
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                  example: "Albert"
                lastName:
                  type: string
                  example: "Saelices"
                email:
                  type: string
                  format: email
                  example: "albert.saelices@gmail.com"
                phone:
                  type: string
                  example: "656874937"
                address:
                  type: string
                  example: "Calle Sant Dalmir 2, 5ºB. Barcelona"
                educations:
                  type: array
                  items:
                    type: object
                    properties:
                      institution:
                        type: string
                        example: "UC3M"
                      title:
                        type: string
                        example: "Computer Science"
                      startDate:
                        type: string
                        format: date
                        example: "2006-12-31"
                      endDate:
                        type: string
                        format: date
                        example: "2010-12-26"
                workExperiences:
                  type: array
                  items:
                    type: object
                    properties:
                      company:
                        type: string
                        example: "Coca Cola"
                      position:
                        type: string
                        example: "SWE"
                      description:
                        type: string
                      startDate:
                        type: string
                        format: date
                        example: "2011-01-13"
                      endDate:
                        type: string
                        format: date
                        example: "2013-01-17"
                cv:
                  type: object
                  properties:
                    filePath:
                      type: string
                      example: "uploads/1715760936750-cv.pdf"
                    fileType:
                      type: string
                      example: "application/pdf"
      responses:
        '201':
          description: Candidato creado exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Candidate added successfully"
                  data:
                    type: object
        '400':
          description: Error en los datos proporcionados
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Error adding candidate"
                  error:
                    type: string

  /candidates/{id}:
    get:
      summary: Obtiene un candidato por su ID
      description: Recupera la información detallada de un candidato específico
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: ID del candidato
      responses:
        '200':
          description: Candidato encontrado exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  firstName:
                    type: string
                  lastName:
                    type: string
                  email:
                    type: string
                  phone:
                    type: string
                  address:
                    type: string
        '400':
          description: ID inválido
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Invalid ID format"
        '404':
          description: Candidato no encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Candidate not found"
        '500':
          description: Error del servidor
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Internal Server Error" 